refactor: rename component to components
