+++ src/components/molecules/BoardFooter.tsx
