+++ src/components/organisms/chart/Mbti.tsx
